# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Portfolio blog JSON API",
        "description": "Portfolio blog API documentation",
        "contact": {
            "email": "yseffk@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/blog/items": {
            "get": {
                "tags": [
                    "Admin Content"
                ],
                "summary": "Get Content Items  whith pagination",
                "description": "Display a listing of the resource.",
                "operationId": "79d603ec9bd2fd1028541854025e2cd3",
                "parameters": [
                    {
                        "name": "pagination",
                        "in": "query",
                        "description": "Need pagination",
                        "required": false,
                        "example": "1"
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "description": "filtering by title ",
                        "required": false,
                        "example": "Lorem.."
                    },
                    {
                        "name": "is_published",
                        "in": "query",
                        "description": "filtering by 'is published' flag",
                        "required": false,
                        "example": "1"
                    },
                    {
                        "name": "is_free",
                        "in": "query",
                        "description": "filtering by 'is free' flag",
                        "required": false,
                        "example": "1"
                    },
                    {
                        "name": "post_id",
                        "in": "query",
                        "description": "filtering by 'Month block item' ID",
                        "required": false,
                        "example": "1"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page",
                        "required": false,
                        "example": "1"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Per page limit",
                        "required": false,
                        "example": "15"
                    },
                    {
                        "name": "order_by",
                        "in": "query",
                        "description": "Sort  list criteria. Allowed fields are: id, sort, created_at. Prepend '-' sign to revert order",
                        "required": false,
                        "example": "-id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin Content"
                ],
                "summary": "Create Content item.",
                "operationId": "ad260c43f2505d2b3077d406f53f647c",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "post_id",
                                    "title",
                                    "is_published",
                                    "sort"
                                ],
                                "properties": {
                                    "post_id": {
                                        "description": "Type item ID",
                                        "type": "integer",
                                        "nullable": false
                                    },
                                    "title": {
                                        "description": "Item title",
                                        "type": "string",
                                        "nullable": false
                                    },
                                    "is_published": {
                                        "description": "Item published flag",
                                        "type": "integer",
                                        "nullable": false
                                    },
                                    "sort": {
                                        "description": "sort",
                                        "type": "integer",
                                        "nullable": false
                                    },
                                    "brief_content": {
                                        "description": "",
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "raw_content": {
                                        "description": "",
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Request validation error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Validation information.",
                                            "type": "string",
                                            "example": "The given data was invalid.",
                                            "nullable": false
                                        },
                                        "errors": {
                                            "description": "Object with all errors.",
                                            "type": "object",
                                            "nullable": false
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/admin/blog/items/{item}": {
            "get": {
                "tags": [
                    "Admin Content"
                ],
                "summary": "Show Content item.",
                "description": "Display the specified resource.",
                "operationId": "2e401b5ec6a35a0e957d628345f319e3",
                "parameters": [
                    {
                        "name": "item",
                        "in": "path",
                        "description": "item ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin Content"
                ],
                "summary": "Delete Content item by ID",
                "operationId": "5eaf877cb50fee6f8ec78891fcad1c2b",
                "parameters": [
                    {
                        "name": "item",
                        "in": "path",
                        "description": "item ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/admin/blog/items/{item}?_method=PUT": {
            "post": {
                "tags": [
                    "Admin Content"
                ],
                "summary": "Update Content item by ID.",
                "operationId": "a61ea8a45389ac6cda0473abacf1021e",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "post_id",
                                    "title",
                                    "is_published",
                                    "sort"
                                ],
                                "properties": {
                                    "post_id": {
                                        "description": "Type item ID",
                                        "type": "integer",
                                        "nullable": false
                                    },
                                    "title": {
                                        "description": "Item title",
                                        "type": "string",
                                        "nullable": false
                                    },
                                    "is_published": {
                                        "description": "Item published flag",
                                        "type": "integer",
                                        "nullable": false
                                    },
                                    "sort": {
                                        "description": "sort",
                                        "type": "integer",
                                        "nullable": false
                                    },
                                    "brief_content": {
                                        "description": "",
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "raw_content": {
                                        "description": "",
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Model not found"
                    },
                    "422": {
                        "description": "Request validation error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Validation information.",
                                            "type": "string",
                                            "example": "The given data was invalid.",
                                            "nullable": false
                                        },
                                        "errors": {
                                            "description": "Object with all errors.",
                                            "type": "object",
                                            "nullable": false
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}